Add special case for text variables in combine-ncc.
This is needed for PPA version of land model, since it introduces a list
of plant species names in its vegn2 restart files. Previous version of
combine-ncc was crashing on text variables because it uses get/put
netcdf calls in the assumption that the variable values can be converted
to double precision numbers and back. This assumption is of course not
valid for text.

Note that this version does not support compressed text variables. If it
detects one, it stops with error message.