Fix gfdl-ws/env.sh with gcc module load
- The reason is totally over my head, but here's the quote from Seth
The reason for this is icc uses the version of gcc it finds on the system, and then behaves like that version.  Since our workstations have 4.8.5 installed at the OS level, icc will (by default) support the same as gcc 4.8.5.  The `_Generic` option was supported in gcc version 4.9.  (Reference https://software.intel.com/content/www/us/en/develop/documentation/cpp-compiler-developer-guide-and-reference/top/compatibility-and-portability/gcc-compatibility-and-interoperability.html)