Update land_utils testing and move to new layout
Updated the land_utils application testing to not require bats.  Included with this is a new testing framework, which was acquired from GNU coreutils.  This testing framework, while not as extensive as bats, has several built-in features that improve the testing process.

This update also begins the migration of the utilities out of the postpocessing and tools directories.